KAN-7 added component 'nav-account' that import and render sidebar-items for authorized/unauthorized user
