+++ src/components/app-sidebar/app-sidebar.tsx
@@ -16,7 +16,6 @@ import {
 
 import { NavMain } from "./nav-main";
 import { NavProjects } from "./nav-projects";
-import { NavUser } from "./nav-user";
 import { TeamSwitcher } from "./team-switcher";
 import {
 	Sidebar,
@@ -26,6 +25,7 @@ import {
 	SidebarRail,
 } from "@/components/ui/sidebar";
 import Link from "next/link";
+import NavAccount from "./nav-account/index";
 
 // This is sample data.
 const data = {
@@ -161,14 +161,14 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
 	return (
 		<Sidebar collapsible="icon" {...props}>
 			<SidebarHeader>
-				<TeamSwitcher teams={data.teams} linkComponent={Link} />
+				<TeamSwitcher teams={data.teams} Link={Link} />
 			</SidebarHeader>
 			<SidebarContent>
-				<NavMain items={data.navMain} linkComponent={Link} />
-				<NavProjects projects={data.projects} linkComponent={Link} />
+				<NavMain items={data.navMain} Link={Link} />
+				<NavProjects projects={data.projects} Link={Link} />
 			</SidebarContent>
 			<SidebarFooter>
-				<NavUser user={data.user} linkComponent={Link} />
+				<NavAccount user={data.user} Link={Link} />
 			</SidebarFooter>
 			<SidebarRail />
 		</Sidebar>
